+++ src/ftxui/dom/dbox.cpp
@@ -4,7 +4,7 @@
 #include <vector>     // for vector
 
 #include "ftxui/dom/elements.hpp"     // for Element, Elements, dbox
-#include "ftxui/dom/node.hpp"         // for Node
+#include "ftxui/dom/node.hpp"         // for Node, Elements
 #include "ftxui/dom/requirement.hpp"  // for Requirement
 #include "ftxui/screen/box.hpp"       // for Box
 
@@ -12,7 +12,7 @@ namespace ftxui {
 
 class DBox : public Node {
  public:
-  DBox(Elements children) : Node(std::move(children)) {}
+  explicit DBox(Elements children) : Node(std::move(children)) {}
 
   void ComputeRequirement() override {
     requirement_.min_x = 0;
@@ -21,6 +21,7 @@ class DBox : public Node {
     requirement_.flex_grow_y = 0;
     requirement_.flex_shrink_x = 0;
     requirement_.flex_shrink_y = 0;
+    requirement_.selection = Requirement::NORMAL;
     for (auto& child : children_) {
       child->ComputeRequirement();
       requirement_.min_x =
@@ -38,8 +39,9 @@ class DBox : public Node {
   void SetBox(Box box) override {
     Node::SetBox(box);
 
-    for (auto& child : children_)
+    for (auto& child : children_) {
       child->SetBox(box);
+    }
   }
 };